Wire the pricing core into the canvas through WASM
Adds @picasso/canvas-pricing: the Rust pricing core loaded and typed for
the browser, and the StrategyNode that computes through it. The package
marshals and never computes — the PRD's bit-identity requirement (7.1)
only holds if the browser runs the same compiled code the server does,
so there is no wasm-bindgen layer and no JavaScript arithmetic.

The grid crosses the boundary once. A 40-leg book over a 25x15 grid is
15,000 repricings; the book is pushed leg by leg, repriced in one call,
and read out of linear memory as a typed array. In Chromium the 40-leg
mixed book runs 12.6ms against a 90ms p95 budget.

Extends verify-wasm-parity.mjs to cover the grid, which immediately
found a cross-target bug scalar parity could not see. All 2250 cells
agreed bit for bit; only the guard's max_error differed (0.124 native,
0.298 WASM). Since the output cells are the fast path, a disagreement
confined to max_error meant the guard sampled different cells — and it
did: Lcg::below reduced with `(x >> 11) as usize % bound`, and usize is
32-bit on wasm32, so the cast dropped 21 bits before the modulo. The
same book carried two different badges and two different cache keys on
client and server. The reduction now happens in u64.

Also:
- pc_implied_vol_reason, so the TS side can tell "below intrinsic" (a
  bad quote) from "not identifiable" (an honest "--").
- pc_american_detail and the grid axes, so a chart labels its cells with
  the values they were priced at rather than a JS reconstruction.
- payoff.html and a Playwright harness. Cells the guard escalated are
  drawn with a dot; on the 40-leg book they form three contiguous
  columns around spot 88-93, the band where the American puts carry
  early-exercise value.
- Widened the demo and parity books to mark odd legs American too. They
  were all calls, and with q below r an American call has no
  early-exercise value, so the guard was measuring against nothing.

31 new tests; 318 across the workspace, plus 20 Rust tests.

+++ crates/pricing-core/src/grid.rs
@@ -165,7 +165,13 @@ impl Lcg {
         if bound == 0 {
             0
         } else {
-            (self.next_u64() >> 11) as usize % bound
+            // The reduction happens in u64, before any narrowing. `as usize`
+            // first would truncate to 32 bits on wasm32 and to 64 natively, so
+            // the client and the server would sample *different cells* — and
+            // the grid would carry two different badges and two different cache
+            // keys for the same book. Found by the parity harness, which is
+            // what it is for.
+            ((self.next_u64() >> 11) % bound as u64) as usize
         }
     }
 }
